+++ src/Components/Oscillator.cpp
@@ -6,17 +6,17 @@
 
 namespace audio
 {
-	Oscillator::Oscillator(const double& dFrequency, const double& dAmplitude)
-		: m_dFrequency(0), m_dAmplitude(0.0), m_dPhase(0.0), m_Wave(Sine)
+	Oscillator::Oscillator(const float& dFrequency, const float& dAmplitude)
+		: m_dFrequency(0.0f), m_dAmplitude(0.0f), m_dPhase(0.0f), m_Wave(Sine)
 	{
 		SetFrequency(dFrequency);
 		SetAmplitude(dAmplitude);
 	}
 
-	void Oscillator::CalcSample(std::array<double, CHANNELS> & dSample)
+	void Oscillator::CalcSample(std::array<float, CHANNELS> & dSample)
 	{
-		const double sine = sin(m_dFrequency * TWO_PI<double> * SampleTime() + (m_dPhase * TWO_PI<double>));
-		double wave = 0.0;
+		const float sine = sin(m_dFrequency * float(TWO_PI<double> * SampleTime()) + (m_dPhase * TWO_PI<float>));
+		float wave = 0.0f;
 		switch (m_Wave)
 		{
 		case audio::Oscillator::Sine:
@@ -26,13 +26,13 @@ namespace audio
 			wave = signbit(sine);
 			break;
 		case audio::Oscillator::Triangle:
-			wave = asin(sine) * TWO_OVER_PI<double>;
+			wave = asin(sine) * TWO_OVER_PI<float>;
 			break;
 		case audio::Oscillator::Saw:
-			wave = -2 / PI<double> * atan(1.0 / tan(m_dFrequency * SampleTime() * PI<double> + (m_dPhase * PI<double>)));
+			wave = -2.0f / PI<float> * atan(1.0f / tan(m_dFrequency * float(SampleTime() * PI<double>) + (m_dPhase * PI<float>)));
 			break;
 		case audio::Oscillator::Noise:
-			wave = 2.0 * ((double)rand() / (double)RAND_MAX) - 1.0;
+			wave = 2.0f * ((float)rand() / (float)RAND_MAX) - 1.0f;
 			break;
 		default:
 			break;
@@ -45,33 +45,33 @@ namespace audio
 		}
 	}
 
-	void Oscillator::SetAmplitude(const double& dNewAmplitude)
+	void Oscillator::SetAmplitude(const float& dNewAmplitude)
 	{
-		m_dAmplitude = Utilities::Clamp(dNewAmplitude, 0.0, 1.0);
+		m_dAmplitude = Utilities::Clamp(dNewAmplitude, 0.0f, 1.0f);
 	}
 
-	const double& Oscillator::GetAmplitude() const
+	const float& Oscillator::GetAmplitude() const
 	{
 		return m_dAmplitude;
 	}
 
-	void Oscillator::SetFrequency(const double& dNewFrequency)
+	void Oscillator::SetFrequency(const float& dNewFrequency)
 	{
-		m_dFrequency = Utilities::Clamp(dNewFrequency, 0.0, 20000.0);
+		m_dFrequency = Utilities::Clamp(dNewFrequency, 0.0f, 20000.0f);
 	}
 
-	const double& Oscillator::GetFrequency() const
+	const float& Oscillator::GetFrequency() const
 	{
 		return m_dFrequency;
 	}
 
-	void Oscillator::SetPhase(const double& dNewPhase)
+	void Oscillator::SetPhase(const float& dNewPhase)
 	{
-		double temp;
+		float temp;
 		m_dPhase = abs(modf(dNewPhase, &temp));
 	}
 
-	const double& Oscillator::GetPhase() const
+	const float& Oscillator::GetPhase() const
 	{
 		return m_dPhase;
 	}
